+++ index.html
@@ -118,8 +118,8 @@
       <a-entity
         position="-2.04 -0.32 -3.81"
         rotation="-14.54 25 0">
-        <a-entity id="scoreboard2"
-          counter="value: 00:00">
+        <a-entity
+          timer-counter="value: 00:00">
         </a-entity>
         <a-entity id="lifes-counter"
           position="0.07 -0.25 -0.01"
@@ -131,8 +131,8 @@
       <a-entity
         position="1.5 -0.32 -4"
         rotation="-14.54 -25 0">
-        <a-entity id="counter234"
-          points="value: 000"
+        <a-entity
+          points-counter="value: 000"
           gamestate-bind="points: points.value">
         </a-entity>
       </a-entity>
