+++ src/components/proxy_event.js
@@ -7,7 +7,6 @@ AFRAME.registerComponent('proxy_event', {
 
   init: function () {
     this.el.sceneEl.addEventListener(this.data.event, function (event) {
-      console.log(event, this.data);
       this.data.dst.emit(this.data.event, event, this.data.bubbles);
     }.bind(this));
   }
