+++ src/index.js
@@ -34,7 +34,9 @@ require('./components/gamestate.js');
 require('./components/gamestate-debug.js');
 require('./components/shoot-controls.js');
 require('./components/bullet.js');
-require('./components/counter.js');
+require('./components/lifes-counter.js');
+require('./components/points-counter.js');
+require('./components/timer-counter.js');
 require('./components/enemy.js');
 require('./components/gun.js');
 require('./components/headset.js');
